Add files via upload
I've Developed in Python, the Weather Application allows users to know the weather of any city in the world as it works with the OpenWeatherMap API. It is an easy to use application with a consistent graphical user interface built using the tkinter framework. Within the app there’s clear instructions on how to get the weather for a specific town by entering the name of the town and view the present temperature, humidity, wind speed and weather satellite images. When you search for a city, app architecture sends an API request, retrieves data as a JSON, and processes it into a neat presentation. A dependable Internet is required for the use of the application